+++ BookNode/server/controllers/userController.js
@@ -1,8 +1,6 @@
 const { userModel, userValidate } = require("../models/userModel");
 const mongoose = require("mongoose");
 const { bookModel } = require("../models/bookModel");
-const jwt = require("jsonwebtoken");
-const bycrypt = require("bcryptjs");
 const { object } = require("joi");
 const SECRET_KEY = process.env.SECRET_KEY || "booksNode2021";
 
@@ -85,116 +83,6 @@ const deleteeUser = async (req, res) => {
     })
   }
 };
-const register = async (req, res) => {
-  try {
-    await userModel.findOne({ email: req.body.email }).then((user) => {
-      if (user) {
-        res.status(400).json({ email: "email already exists" });
-      } else {
-        const { firstName, lastName, email, password } = req.body;
-        const newUser = new userModel({
-          firstName: firstName,
-          lastName: lastName,
-          email: email,
-          password: password,
-        });
-        bycrypt.genSalt(10, (err, salt) => {
-          bycrypt.hash(newUser.password, salt, (error, hash) => {
-            if (error) throw error;
-            newUser.password = hash;
-            newUser
-              .save()
-              .then((data) => res.json(data))
-              .catch((err) => {
-                console.log("someting wrong");
-              });
-          });
-        });
-      }
-    });
-  } catch (err) {
-    res.status(301).json({
-      message: "error in register",
-      success: false,
-      error: err.message
-    })
-  }
-};
-
-const logIn = async (req, res) => {
-  // בודקים שהסיסמה והאיימל שקיבלנו מהקליינט הכן קיימים
-  const { email, password } = req.body.user;
-  if (!email || !password) {
-    return res.status(400).json({
-      success: false,
-      message: "there is error with email or password.",
-      error: "email and password is required ",
-    });
-  }
-
-  // מתחילים ניסוי של מצאית משתמש על פי הנתונים מהקליינט ומגבים בהתאם
-  try {
-    // מנסים למצוא משתמש לפי איימל
-    const user = await userModel.findOne({ email });
-
-    // במידה ולא קיים מחזירים תשובה לקליינט שלא קיים משתמש כזה
-    if (!user) {
-      return res.status(400).json({
-        success: false,
-        message: "email not found",
-        errors: { email: "email not fond" },
-      });
-    }
-
-    /*
-           אין 
-           else
-           בגלל שבמידה והתנאי מתקיים אנחנו יוצאים מהפונקציה מכיוון שהחזרנו תשובה לקליינט 
-    */
-
-    // משווים את הסיסמה שהקיבלנו מהקליינט לסיסמה של המשתמש שמצאנו על פי המייל ותופסים במשתנה את מה שחוזר לנו מהפונקציה של ההשוואה
-    const isPasswordCorrect = await bycrypt.compare(password, user.password);
-
-    // במידה ומה שחזר מהפונקציה של ההשוואה לא תואם אז מחזירים לקליינט תשובה שהסיסמה שגויה ולא מתאימה
-    if (!isPasswordCorrect) {
-      return res.status(400).json({
-        success: false,
-        message: "wrong password",
-        errors: { password: "wrong password" },
-      });
-    }
-
-    /*
-           אין 
-           else
-           בגלל שבמידה והתנאי מתקיים אנחנו יוצאים מהפונקציה מכיוון שהחזרנו תשובה לקליינט 
-    */
-
-    // מוחקים את הסיסמה של המשתמש כדי שנוכל להחזיר לקליינט נתונים רלוונטים ולא רגישים מדי
-    delete user.password;
-
-    /* 
-
-        יוצרים תוקן להחזיר לקליינט במידה והכל עבר בהצלחה 
-        זה ישמש אותנו בהמשך כדי לאמת נתונים של משתמש לפני שנציג לנו מידע מסוים 
-        זה יעזור לנו לוודא שהכן הבקשות שנשלחות מהקליינט הכן אמינות
-
-    */
-    const token = jwt.sign(user.toJSON(), SECRET_KEY, { expiresIn: "1d" });
-    res.status(200).json({
-      success: true,
-      message: "success",
-      data: token,
-    });
-
-    // במידה ובמהלך ניסיון ההרצה של הפונקציה היא נכשלת או נזרקת שגיעה היא תתפס כאן
-  } catch (err) {
-    res.status(500).json({
-      message: "something went wrong",
-      error: err.message,
-    });
-  }
-};
 
 const addBookToUser = async (req, res) => {
   try {
@@ -234,7 +122,5 @@ module.exports = {
   updateUser,
   deleteeUser,
   addBookToUser,
-  register,
-  logIn,
   showBooks,
 };
