+++ BookNode/server/middleware/isToken.js
@@ -1,8 +1,10 @@
 const jwt = require('jsonwebtoken');
 const SECRET_KEY = process.env.SECRET_KEY || "booksNode2021";;
 const checkAuth = (req, res, next) => {
+    const headers = req.headers("Authorization")
     try {
         // const token = req.headers['authorization'].split(' ')[1];
+        const token = headers.split(' ')[1];
         jwt.verify(token, SECRET_KEY)
         next();
     }
